+++ src/backend_task/wallet/sign_message_with_key.rs
@@ -3,26 +3,13 @@
 
 use crate::backend_task::BackendTaskSuccessResult;
 use crate::backend_task::error::TaskError;
+use crate::backend_task::wallet::dash_signed_message;
 use crate::context::AppContext;
 use crate::model::wallet::WalletSeedHash;
-use dash_sdk::dpp::dashcore::hashes::Hash;
-use dash_sdk::dpp::dashcore::secp256k1::{Message, Secp256k1, SecretKey};
-use dash_sdk::dpp::dashcore::sign_message::{MessageSignature, signed_msg_hash};
+use dash_sdk::dpp::dashcore::secp256k1::SecretKey;
 use dash_sdk::dpp::identity::KeyType;
 use dash_sdk::dpp::key_wallet::bip32::DerivationPath;
 use std::sync::Arc;
-
-/// Build the Base64-encoded Dash signed-message envelope for `message` signed
-/// with `secret_key`. The envelope is a recoverable signature: a header byte
-/// (`27 + recId`, `+4` when `compressed`) followed by the 64-byte signature, so
-/// a verifier can recover the signer's public key from the signature alone.
-fn dash_signed_message(message: &str, secret_key: &SecretKey, compressed: bool) -> String {
-    let secp = Secp256k1::new();
-    let message_hash = signed_msg_hash(message);
-    let digest = Message::from_digest(*message_hash.as_byte_array());
-    let recoverable = secp.sign_ecdsa_recoverable(&digest, secret_key);
-    MessageSignature::new(recoverable, compressed).to_base64()
-}
 
 impl AppContext {
     /// Sign a message with a wallet-derived key at `derivation_path`.
@@ -96,39 +83,3 @@ impl AppContext {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use dash_sdk::dpp::dashcore::secp256k1::PublicKey;
-    use dash_sdk::dpp::dashcore::sign_message::signed_msg_hash;
-
-    /// The envelope round-trips: the signer's public key recovers from the
-    /// produced signature for both compression flags. A hardcoded recovery
-    /// header would fail ~50% of the time here.
-    fn assert_recovers(compressed: bool) {
-        let secp = Secp256k1::new();
-        let secret_key = SecretKey::from_byte_array(&[0x42u8; 32]).expect("valid secret");
-        let expected_pubkey = PublicKey::from_secret_key(&secp, &secret_key);
-        let message = "Bilby was here";
-
-        let base64 = dash_signed_message(message, &secret_key, compressed);
-        let parsed = MessageSignature::from_base64(&base64).expect("valid envelope");
-        assert_eq!(parsed.compressed, compressed);
-
-        let recovered = parsed
-            .recover_pubkey(&secp, signed_msg_hash(message))
-            .expect("recovers a public key");
-        assert_eq!(recovered.inner, expected_pubkey);
-        assert_eq!(recovered.compressed, compressed);
-    }
-
-    #[test]
-    fn recovers_signer_pubkey_compressed() {
-        assert_recovers(true);
-    }
-
-    #[test]
-    fn recovers_signer_pubkey_uncompressed() {
-        assert_recovers(false);
-    }
-}
